requirements.txt
pyrealsense2
numpy
opencv-python
pyglet
glfw
PyOpenGL